refactor kubelet and eks_fargate checks and put shared code in datadog_checks_base
