+++ kubelet/datadog_checks/kubelet/__init__.py
@@ -1,13 +1,11 @@
 from .__about__ import __version__
-from .common import KubeletCredentials, PodListUtils, get_pod_by_uid, is_static_pending_pod, urljoin
+from .common import PodListUtils, get_pod_by_uid, is_static_pending_pod
 from .kubelet import KubeletCheck
 
 __all__ = [
     'KubeletCheck',
     '__version__',
     'PodListUtils',
-    'KubeletCredentials',
-    'urljoin',
     'get_pod_by_uid',
     'is_static_pending_pod',
 ]
