+++ kubelet/tests/test_common.py
@@ -8,8 +8,9 @@ import sys
 import mock
 import pytest
 
+from datadog_checks.base.checks.kubelet_base.base import KubeletCredentials, urljoin
 from datadog_checks.base.checks.openmetrics import OpenMetricsBaseCheck
-from datadog_checks.kubelet import KubeletCredentials, PodListUtils, get_pod_by_uid, is_static_pending_pod, urljoin
+from datadog_checks.kubelet import PodListUtils, get_pod_by_uid, is_static_pending_pod
 
 from .test_kubelet import mock_from_file
